@@ -27,23 +27,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
-    "@angular/cdk": "^13.3.0",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
-    "@angular/material": "13.3.5",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
+    "@angular/animations": "14.0.7",
+    "@angular/cdk": "16.1.8",
+    "@angular/common": "14.0.7",
+    "@angular/compiler": "14.0.7",
+    "@angular/core": "16.1.9",
+    "@angular/forms": "14.0.7",
+    "@angular/material": "16.1.8",
+    "@angular/platform-browser": "14.0.7",
+    "@angular/platform-browser-dynamic": "14.0.7",
+    "@angular/router": "14.0.7",
+    "@nestjs/common": "10.0.2",
+    "@nestjs/core": "10.0.2",
     "@nestjs/mapped-types": "^2.0.2",
-    "@nestjs/platform-express": "^8.0.0",
+    "@nestjs/platform-express": "10.0.2",
     "@nestjs/typeorm": "^10.0.0",
-    "@nrwl/angular": "13.9.3",
-    "@nrwl/nest": "^13.9.5",
     "chart.js": "^2.9.3",
     "discord.js": "^12.3.1",
     "express": "^4.17.3",
@@ -52,61 +50,66 @@
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.34",
-    "ng2-charts": "^2.2.3",
+    "ng2-charts": "^5.0.2",
     "ngx-cookie-service": "^13.1.2",
     "ngx-quill": "^16.2.1",
     "pg": "^8.11.2",
     "quill": "^1.3.7",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.4.0",
+    "rxjs": "7.8.1",
     "typeorm": "^0.3.17",
     "uuid": "^8.3.2",
     "ws": "^8.5.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "0.13.1",
+    "@nx/angular": "16.6.0",
+    "@nx/nest": "^13.9.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.0",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.2.0",
-    "@angular/compiler-cli": "~13.2.0",
-    "@angular/language-service": "~13.2.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "13.9.3",
-    "@nrwl/cypress": "13.9.3",
-    "@nrwl/eslint-plugin-nx": "13.9.3",
-    "@nrwl/jest": "13.9.3",
-    "@nrwl/linter": "13.9.3",
-    "@nrwl/nest": "13.9.5",
-    "@nrwl/node": "13.9.5",
-    "@nrwl/workspace": "13.9.3",
+    "@angular-devkit/build-angular": "16.1.8",
+    "@angular-devkit/core": "16.1.8",
+    "@angular-devkit/schematics": "16.1.8",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "~16.1.0",
+    "@angular/compiler-cli": "14.0.7",
+    "@angular/language-service": "14.0.7",
+    "@nestjs/schematics": "10.0.1",
+    "@nestjs/testing": "10.0.2",
+    "@nx/workspace": "16.6.0",
+    "@schematics/angular": "16.1.8",
     "@types/faker": "^5.5.9",
-    "@types/jest": "^27.0.2",
+    "@types/jest": "29.4.4",
     "@types/lodash": "^4.14.180",
-    "@types/node": "16.11.7",
+    "@types/node": "18.7.1",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~5.10.0",
+    "@typescript-eslint/eslint-plugin": "5.60.1",
+    "@typescript-eslint/parser": "5.60.1",
     "concurrently": "^7.0.0",
     "cross-env": "^7.0.3",
-    "cypress": "^9.1.0",
-    "eslint": "~8.7.0",
+    "cypress": "^12.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "faker": "^5.5.3",
     "husky": "^8.0.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.1.1",
-    "nx": "13.9.3",
-    "prettier": "^2.5.1",
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "28.1.3",
+    "jest-preset-angular": "13.1.1",
+    "nx": "16.6.0",
+    "prettier": "2.6.2",
     "superstatic": "^7.1.0",
-    "ts-jest": "27.0.5",
+    "ts-jest": "29.1.1",
     "ts-mockito": "^2.6.1",
-    "ts-node": "^10.7.0",
+    "ts-node": "10.9.1",
     "tslib": "^2.4.1",
-    "typescript": "~4.5.2",
-    "wait-on": "^6.0.1"
+    "typescript": "5.1.6",
+    "wait-on": "^6.0.1",
+    "@nx/cypress": "16.6.0",
+    "@nx/eslint-plugin": "16.6.0",
+    "@nx/jest": "16.6.0",
+    "@nx/linter": "16.6.0",
+    "@nx/nest": "16.6.0",
+    "@nx/node": "16.6.0"
   }
 }
