@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, Router } from '@angular/router';
+import { Router } from '@angular/router';
 import { filter, map, Observable, tap } from 'rxjs';
 import { isNonNull } from '../../../shared/helpers';
 import { UserService } from '~shared/services/user-service/user.service';
@@ -7,7 +7,7 @@ import { UserInterface } from '~shared/interfaces/user.interface';
 import { UserAccess } from '~shared/enums/user-access.enum';
 
 @Injectable()
-export class HasAdminRights implements CanActivate {
+export class HasAdminRights  {
   constructor(private router: Router, private userService: UserService) {}
 
   canActivate(): Observable<boolean> {
